requirements.txt
Mastodon.py
tweepy
